Simplified serial parsing for USB.
This separates the USB and serial parsing from the "CONNECTED"
state machine, and adds buffered reads to the received USB commands
to prevent command drops when the buffer is filled up.

This PR also removes the old winch code, because I don't think it
is up to date and I didn't want to port it.